half working progress

+++ PowerPlay/PowerPlay/Models/UserModel.swift
@@ -9,11 +9,13 @@ struct userInfo: Identifiable {
 
 struct ArduinoData: Decodable {
     //let unix_timestamp: Int
-    let totalEnergy: Int
+    let totalEnergy: Float
+    //let unique_id: String
     
     init() {
         //self.unix_timestamp = 200
         self.totalEnergy = 200
+        //self.unique_id = "uniqe id"
     }
     
 }
